requirements.txt
click==7.1.2
Flask==1.1.2
glitch-this==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
Pillow==7.2.0
Werkzeug==1.0.1
